// event-planner-frontend/src/app/core/models/auth.model.ts
import { User } from './user.model';

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  password_confirmation: string;
  phone?: string;
}

export interface LoginResponse {
  success: boolean;
  message: string;
  user: User;
  token: string;
}

export interface AuthResponse {
  success: boolean;
  message: string;
  user: User;
  token: string;
}